+++ components/Hero.module.scss
@@ -1,9 +1,11 @@
+@import "../styles/colors.module.scss";
+
 $text-color: rgba(0, 168, 232, 0.8);
 $bg-color: rgba(0, 168, 232, 0.075);
 
 .hero {
-    color: $text-color;    
-    font-weight: 500;
+    color: $sections-grad-color;    
+    font-weight: 600;
     letter-spacing: 0.2em;
     background-color: $bg-color;
 }
