+++ components/Profession.module.scss
@@ -1,7 +1,9 @@
 @import "../styles/colors.module.scss";
 
+
 .developer {
     letter-spacing: 0.5em;
     color: $sections-grad-color;
     text-transform: uppercase;
+    font-weight: 300;
 }
