+++ components/ProfilePic.tsx
@@ -1,4 +1,4 @@
-import { Center } from "@chakra-ui/react";
+import { Box, Center, Spacer } from "@chakra-ui/react";
 import { motion } from "framer-motion";
 import Image from "next/image";
 import React from "react";
@@ -7,8 +7,12 @@ type Props = {};
 
 function ProfilePic({}: Props) {
   return (
-    <Center
+    <Box
       as={motion.div}
+      w={[192, 240, 300]}
+      mx="auto"
+      mt={[2.5,5]}
+      mb={[5, 10]}
       initial={{ opacity: 0, scale: 0.5 }}
       animate={{ opacity: 1, scale: 1 }}
       whileHover={{ scale: 1.05 }}
@@ -16,9 +20,11 @@ function ProfilePic({}: Props) {
       <Image
         alt="profile"
         src="/lucas-parma-profile.png"
+        layout="responsive"
         width="192"
         height="192"></Image>
-    </Center>
+        <Spacer />
+    </Box>
   );
 }
