+++ container/About/About.module.scss
@@ -8,7 +8,7 @@
 }
 
 .about_text {
-    color: $light-color;
+    color: $sections-grad-color;
     font-weight: 300;
     text-align: center;
     letter-spacing: 0.01em;
