+++ styles/colors.module.scss
@@ -1,15 +1,13 @@
 $dark-color: rgb(21, 15, 22.5);
 $white-color: rgba(227, 230, 219, 1);
 $light-color: rgba(214, 253, 255, 1);
+
 $text-color: rgba(0, 168, 232, 0.6);
 $text-color-green: rgb(220, 255, 121);
-$blue-color: rgba(35, 87, 137, 0.75);
-$sections-color: rgba(35, 87, 137, 0.9);
-$sections-grad-color: rgba(214, 253, 255, 1);
 
-
+$shadow-color: rgba(35, 87, 137, 0.75);
+$sections-grad-color: rgba(174, 205, 234, 1);
 $background-gradient: radial-gradient(circle, hsla(278, 22%, 7%, 1) 0%, hsla(214, 64%, 29%, 1) 50%, hsla(285, 21%, 7%, 1) 100%);
-
 
 $test-color: rgb(204, 255, 51);
 $test-color2: rgba(184, 153, 185, 0.9);
